Move function models back in function model file
