+++ packages/postgres-database/src/simcore_postgres_database/models/funcapi_function_jobs_table.py
@@ -6,36 +6,12 @@
 import uuid
 
 import sqlalchemy as sa
-from models_library.api_schemas_webserver.functions_wb_schema import (
-    FunctionClass,
-    FunctionID,
-    FunctionInputs,
-    FunctionJobClassSpecificData,
-    FunctionJobID,
-    FunctionOutputs,
-)
-from pydantic import BaseModel
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.sql import func
 
 from ._common import RefActions
 from .base import metadata
 from .funcapi_functions_table import functions_table
-
-
-class FunctionJobDB(BaseModel):
-    function_uuid: FunctionID
-    title: str = ""
-    description: str = ""
-    inputs: FunctionInputs
-    outputs: FunctionOutputs
-    class_specific_data: FunctionJobClassSpecificData
-    function_class: FunctionClass
-
-
-class RegisteredFunctionJobDB(FunctionJobDB):
-    uuid: FunctionJobID
-
 
 function_jobs_table = sa.Table(
     "funcapi_function_jobs",
