+++ services/web/server/src/simcore_service_webserver/functions/_functions_repository.py
@@ -13,6 +13,8 @@ from models_library.api_schemas_webserver.functions_wb_schema import (
     FunctionJobIDNotFoundError,
     FunctionOutputs,
     FunctionOutputSchema,
+    RegisteredFunctionDB,
+    RegisteredFunctionJobDB,
 )
 from models_library.rest_pagination import PageMetaInfoLimitOffset
 from simcore_postgres_database.models.funcapi_function_job_collections_table import (
@@ -23,13 +25,9 @@ from simcore_postgres_database.models.funcapi_function_job_collections_to_functi
     function_job_collections_to_function_jobs_table,
 )
 from simcore_postgres_database.models.funcapi_function_jobs_table import (
-    RegisteredFunctionJobDB,
     function_jobs_table,
 )
-from simcore_postgres_database.models.funcapi_functions_table import (
-    RegisteredFunctionDB,
-    functions_table,
-)
+from simcore_postgres_database.models.funcapi_functions_table import functions_table
 from simcore_postgres_database.utils_repos import (
     get_columns_from_db_model,
     transaction_context,
